+++ lgca/ib_base.py
@@ -103,8 +103,6 @@ class IBLGCA_base(LGCA_base, ABC):
         Flux for each possible channel configuration. Dimensions: ``(lgca.K + 1, len(lgca.c))``.
     K : int
         Number of channels per node. Equal to ``lgca.velocitychannels + lgca.restchannels``.
-    n_crit
-        Internal variable for the wetting interaction.
     n_t : :py:class:`numpy.ndarray`
         Sum of particles in the lattice for all timesteps in the previous simulation.
         Only available after a simulation performed with ``timeevo(recordN=True)``. Dimensions: ``(timesteps,)``.